Allow partial LightState objects.
This allows creating states which only affect certain channels. The "off"
state can now simply set brightness to zero, and the color will be fetched
from any state it's blended with, or simply left untouched from whatever
it is already set to.

This change also makes it possible to declare states inline everywhere
states are referenced in the config file. Sometimes declaring a named
state is overkill for stuff like simple overrides, so this is useful.